requirements.txt
slackclient
py-trello
pandas

